fix drag and drop handler
