+++ packages/editor/src/components/EditorLayout.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React from "react";
-import { DndContext, closestCenter } from "@dnd-kit/core";
+import { DndContext, closestCenter, DragEndEvent } from "@dnd-kit/core";
 import { useEditorStore } from "../lib/store";
 import { widgetRegistry } from "../lib/widgetRegistry";
 import { makeOnDragEnd } from "../dnd/handlers";
@@ -29,16 +29,46 @@ function createFromWidget(key: string) {
 export default function EditorLayout() {
   const moveNode = useEditorStore((s) => s.moveNode);
   const addChild = useEditorStore((s) => s.addChild);
+  const page = useEditorStore((s) => s.page);
 
-  const onDragEnd = makeOnDragEnd({ moveNode, addChild, createFromWidget });
+  const canvasDragEnd = makeOnDragEnd({ moveNode, addChild, createFromWidget });
+
+  const handleDragEnd = (event: DragEndEvent) => {
+    const { over, active } = event;
+    if (!over) return;
+    const overId = String(over.id);
+
+    // Canvas drop slots use ids like "drop:<parentId>:<index>"
+    if (overId.startsWith("drop:")) {
+      canvasDragEnd(event);
+      return;
+    }
+
+    // Layers panel drop zones use ids like "drop-<id>"
+    const activeId = String(active.id);
+    const dropZone =
+      overId.startsWith("drop-") && over.data.current?.isContainer;
+
+    if (dropZone) {
+      const newParentId = over.data.current?.parentId as string;
+      const insertIndex =
+        overId === "drop-root" ? (page.children?.length || 0) : 0;
+      moveNode(activeId, newParentId, insertIndex);
+      return;
+    }
+
+    const newParentId = over.data.current?.parentId as string;
+    const overIndex = over?.data?.current?.sortable?.index ?? 0;
+    moveNode(activeId, newParentId, overIndex);
+  };
 
   return (
-    <DndContext collisionDetection={closestCenter} onDragEnd={onDragEnd}>
+    <DndContext collisionDetection={closestCenter} onDragEnd={handleDragEnd}>
       <div className="flex h-screen">
-      <Sidebar />
-      <EditorCanvas />
-      <LayersPanel />
-      <PropertyPanel />
+        <Sidebar />
+        <EditorCanvas />
+        <LayersPanel />
+        <PropertyPanel />
       </div>
     </DndContext>
   );
