+++ packages/editor/tsconfig.json
@@ -5,8 +5,7 @@
     "declaration": true,
     "jsx": "react-jsx",
     "types": ["react", "react-dom"],
-    "skipLibCheck": true,
-    "paths": {}
+    "skipLibCheck": true
   },
   "include": ["src/**/*"]
 }
